Initial spring security

+++ src/main/java/com/empowerpeers/entity/CourseRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
  * Basic Crud Repository for Courses
  * @author Dave
  */
-public interface CourseRepository extends CrudRepository<Course, Long> {
+public interface CourseRepository extends CrudRepository<Course, Integer> {
 }
